fix: resolve syntax errors and enum method issues in Phase 1 migration refactoring
- Fix corrupted syntax in proposals migration line 28 (constrained('s('science_clusters') -> constrained('science_clusters'))
- Simplify MigrationHelpers::addCheckConstraint to accept string table names only (CHECK constraints must be after table creation)
- Add SQLite compatibility handling (skip ALTER TABLE ADD CONSTRAINT on SQLite)
- Add missing values() static method to ReviewRecommendation, OutputStatusType, and AuthorStatus enums
- Make addCheckConstraintToTable delegate to addCheckConstraint for DRY
- Update enumToStringWithCheck to document that CHECK constraints must be added outside Schema callbacks
- All 242 tests pass (13 skipped, 1 risky)

+++ database/migrations/2025_10_26_111048_create_science_clusters_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('science_clusters', function (Blueprint $table) {
             $table->id();
             $table->foreignId('parent_id')->nullable()->comment('Parent Rumpun Ilmu (Self-reference)')->constrained('science_clusters')->onDelete('cascade');
-            $table->tinyInteger('level')->comment('Level 1, 2, or 3');
+            $table->smallInteger('level')->comment('Level 1, 2, or 3');
             $table->string('name')->comment('Nama Rumpun Ilmu');
             $table->timestamps();
         });
